Update config.yml
Fixed "typo" that was pointed out on dev group. From prior experience this did not actually prevent anything from working. I double checked this morning and during start up on my server it listed the spawn restriction for the zombie pigman but i figured might as well change it just for consistencies sake.